requirements.txt
google-cloud-bigquery==3.16.0
python-decouple==3.8
paramiko==3.4.0
pyarrow==14.0.2 
pandas==2.1.4
click==8.1.7